Added (L)Ubuntu boot that sort of works
"mount" does nothing, but running it manually still works and a VM with 512 MB of RAM can boot the system... into install mode.
There's apparently no way to get the live CD version, any info on the internet is from around 2011 and completely and utterly outdated,
so in the last 7 years nobody attempted to boot Ubuntu over PXE or what?